# backend/openapi.yaml
components:
...
  schemas:
...
    AddRule:
      properties:
        compensatory_measures:
          nullable: true
          type: string
        criteria:
          items:
            type: object
          type: array
        name:
          type: string
        protections:
          items:
            $ref: '#/components/schemas/RuleProtection'
          type: array
      type: object
...
    BinaryResponse:
      properties:
        message:
          example: Failed load data
          type: string
        result:
          example: false
          type: boolean
      type: object
...
    PaginationResponse:
      properties:
        limit:
          example: 10
          type: integer
        page:
          example: 1
          type: integer
        total_items:
          example: 10
          type: integer
        total_pages:
          example: 1
          type: integer
      type: object
...
    Question1:
      properties:
        answers:
          items:
            $ref: '#/components/schemas/QuestionAnswer'
          type: array
        id:
          type: integer
        id_right_answer:
          nullable: true
          type: integer
        number_question:
          nullable: true
          type: integer
        text:
          type: string
      type: object
...
    QuestionList:
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
        questions:
          items:
            $ref: '#/components/schemas/Question1'
          type: array
      type: object
...
    Rule:
      properties:
        compensatory_measures:
          nullable: true
          type: string
        criteria:
          items:
            $ref: '#/components/schemas/Criteria'
          type: array
        id:
          type: integer
        name:
          type: string
        protections:
          items:
            $ref: '#/components/schemas/Protection'
          type: array
        questions:
          items:
            type: object
          type: array
      type: object
    RuleProtection:
      properties:
        id:
          type: integer
        is_demasking:
          type: boolean
        is_masking:
          type: boolean
      type: object
...
    UpdateRule:
      properties:
        compensatory_measures:
          nullable: true
          type: string
        criteria:
          items:
            type: object
          type: array
        id:
          type: integer
        name:
          type: string
        protections:
          items:
            $ref: '#/components/schemas/RuleProtection'
          type: array
      type: object
...
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
    roleAuth:
      in: header
      name: Role
      type: apiKey
    userAuth:
      in: header
      name: User
      type: apiKey
info:
  title: "\u041F\u0440\u0438\u043B\u043E\u0436\u0435\u043D\u0438\u0435 \u0434\u043B\
    \u044F \u0433\u0435\u043D\u0435\u0440\u0430\u0446\u0438\u0438 \u043A\u0430\u0440\
    \u0442 \u043C\u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\u0438\u044F"
  version: v0.1
openapi: 3.0.3
paths:
...
  /api/rule/rule/:
    post:
      description: "\u0414\u043E\u0431\u0430\u0432\u043B\u0435\u043D\u0438\u0435 \u043F\
        \u0440\u0430\u0432\u0438\u043B\u0430"
      requestBody:
        content:
          app/json:
            schema:
              $ref: '#/components/schemas/AddRule'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: null
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: null
      summary: "\u0414\u043E\u0431\u0430\u0432\u0438\u0442\u044C \u043F\u0440\u0430\
        \u0432\u0438\u043B\u043E \u0444\u043E\u0440\u043C\u0438\u0440\u043E\u0432\u0430\
        \u043D\u0438\u044F \u043A\u0430\u0440\u0442 \u043C\u0430\u0441\u043A\u0438\
        \u0440\u043E\u0432\u0430\u043D\u0438\u044F"
      tags:
      - rule
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0435\u043D\u0438\u0435 \u043F\u0440\
        \u0430\u0432\u0438\u043B"
      parameters:
      - in: query
        name: rule_id
        required: false
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
          description: "\u0421\u043F\u0438\u0441\u043E\u043A \u043E\u0442\u0434\u0435\
            \u043B\u043E\u0432"
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u041E\u0448\u0438\u0431\u043A\u0430 \u043F\u0440\u0438 \u0432\
            \u044B\u043F\u043E\u043B\u043D\u0435\u043D\u0438\u0438 \u0437\u0430\u043F\
            \u0440\u043E\u0441\u0430"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u043F\u0440\u0430\
        \u0432\u0438\u043B\u043E \u0444\u043E\u0440\u043C\u0438\u0440\u043E\u0432\u0430\
        \u043D\u0438\u044F \u043A\u0430\u0440\u0442 \u043C\u0430\u0441\u043A\u0438\
        \u0440\u043E\u0432\u0430\u043D\u0438\u044F"
      tags:
      - rule
    put:
      description: "\u0418\u0437\u043C\u0435\u043D\u0435\u043D\u0438\u0435 \u043F\u0440\
        \u0430\u0432\u0438\u043B\u0430"
      requestBody:
        content:
          app/json:
            schema:
              $ref: '#/components/schemas/UpdateRule'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: null
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: null
      summary: "\u0418\u0437\u043C\u0435\u043D\u0438\u0442\u044C \u043F\u0440\u0430\
        \u0432\u0438\u043B\u043E \u0444\u043E\u0440\u043C\u0438\u0440\u043E\u0432\u0430\
        \u043D\u0438\u044F \u043A\u0430\u0440\u0442 \u043C\u0430\u0441\u043A\u0438\
        \u0440\u043E\u0432\u0430\u043D\u0438\u044F"
      tags:
      - rule
    delete:
      description: "\u0423\u0434\u0430\u043B\u0435\u043D\u0438\u0435 \u043F\u0440\u0430\
        \u0432\u0438\u043B\u0430"
      parameters:
      - in: query
        name: rule_id
        type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: null
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: null
      summary: "\u0423\u0434\u0430\u043B\u0438\u0442\u044C \u043F\u0440\u0430\u0432\
        \u0438\u043B\u043E"
      tags:
      - rule
  /api/rule/rules/:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u043F\u0438\
        \u0441\u043E\u043A \u043F\u0440\u0430\u0432\u0438\u043B"
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 1
          example: 1
          minimum: 0
          type: integer
      - explode: true
        in: query
        name: protection_ids[]
        required: false
        schema:
          items:
            type: integer
          type: array
        style: form
      - explode: true
        in: query
        name: rule_ids[]
        required: false
        schema:
          items:
            type: integer
          type: array
        style: form
      - in: query
        name: name
        required: false
        schema:
          nullable: true
          type: string
      - explode: true
        in: query
        name: type_location_ids[]
        required: false
        schema:
          items:
            type: integer
          type: array
        style: form
      - in: query
        name: limit
        required: false
        schema:
          default: 1000
          example: 10
          type: integer
      - explode: true
        in: query
        name: type_work_ids[]
        required: false
        schema:
          items:
            type: integer
          type: array
        style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Rule'
                type: array
          description: null
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u041E\u0448\u0438\u0431\u043A\u0430 \u043F\u0440\u0438 \u0432\
            \u044B\u043F\u043E\u043B\u043D\u0435\u043D\u0438\u0438 \u0437\u0430\u043F\
            \u0440\u043E\u0441\u0430"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u043F\u0438\
        \u0441\u043E\u043A \u043F\u0440\u0430\u0432\u0438\u043B"
      tags:
      - rule
  /api/rule/questions/:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u043F\u0438\
        \u0441\u043E\u043A \u0432\u043E\u043F\u0440\u043E\u0441\u043E\u0432 \u0434\
        \u043B\u044F \u0443\u0442\u043E\u0447\u043D\u0435\u043D\u0438\u044F \u043C\
        \u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\u0438\u044F \u0437\u0430\
        \u0449\u0438\u0442"
      parameters:
      - explode: true
        in: query
        name: location_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      - in: query
        name: page
        required: false
        schema:
          default: 1
          example: 1
          minimum: 0
          type: integer
      - explode: true
        in: query
        name: question_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      - in: query
        name: text
        required: false
        schema:
          nullable: true
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 1000
          example: 10
          type: integer
      - explode: true
        in: query
        name: type_work_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
          description: null
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u041E\u0448\u0438\u0431\u043A\u0430 \u043F\u0440\u0438 \u0432\
            \u044B\u043F\u043E\u043B\u043D\u0435\u043D\u0438\u0438 \u0437\u0430\u043F\
            \u0440\u043E\u0441\u0430"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u043F\u0438\
        \u0441\u043E\u043A \u0432\u043E\u043F\u0440\u043E\u0441\u043E\u0432 \u0434\
        \u043B\u044F \u0443\u0442\u043E\u0447\u043D\u0435\u043D\u0438\u044F \u043C\
        \u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\u0438\u044F \u0437\u0430\
        \u0449\u0438\u0442"
      tags:
      - rule
  /api/rule/filter-question-rule/:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u043F\u0438\
        \u0441\u043E\u043A \u0432\u043E\u043F\u0440\u043E\u0441\u043E\u0432 \u0434\
        \u043B\u044F \u0443\u0442\u043E\u0447\u043D\u0435\u043D\u0438\u044F \u043C\
        \u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\u0438\u044F \u0437\u0430\
        \u0449\u0438\u0442"
      parameters:
      - explode: true
        in: query
        name: location_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      - in: query
        name: page
        required: false
        schema:
          default: 1
          example: 1
          minimum: 0
          type: integer
      - explode: true
        in: query
        name: question_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      - in: query
        name: text
        required: false
        schema:
          nullable: true
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 1000
          example: 10
          type: integer
      - explode: true
        in: query
        name: type_work_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
          description: null
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u041E\u0448\u0438\u0431\u043A\u0430 \u043F\u0440\u0438 \u0432\
            \u044B\u043F\u043E\u043B\u043D\u0435\u043D\u0438\u0438 \u0437\u0430\u043F\
            \u0440\u043E\u0441\u0430"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0441\u043F\u0438\
        \u0441\u043E\u043A \u0432\u043E\u043F\u0440\u043E\u0441\u043E\u0432 \u0434\
        \u043B\u044F \u0443\u0442\u043E\u0447\u043D\u0435\u043D\u0438\u044F \u043C\
        \u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\u0438\u044F \u0437\u0430\
        \u0449\u0438\u0442"
      tags:
      - rule
  /api/rule/question/:
    get:
      description: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0432\u043E\u043F\
        \u0440\u043E\u0441 \u0434\u043B\u044F \u0443\u0442\u043E\u0447\u043D\u0435\
        \u043D\u0438\u044F \u043C\u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\
        \u0438\u044F \u0437\u0430\u0449\u0438\u0442"
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question1'
          description: null
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u041E\u0448\u0438\u0431\u043A\u0430 \u043F\u0440\u0438 \u0432\
            \u044B\u043F\u043E\u043B\u043D\u0435\u043D\u0438\u0438 \u0437\u0430\u043F\
            \u0440\u043E\u0441\u0430"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0432\u043E\u043F\
        \u0440\u043E\u0441 \u0434\u043B\u044F \u0443\u0442\u043E\u0447\u043D\u0435\
        \u043D\u0438\u044F \u043C\u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\
        \u0438\u044F \u0437\u0430\u0449\u0438\u0442"
      tags:
      - rule
...
servers:
- description: "\u0422\u0435\u0441\u0442\u043E\u0432\u044B\u0439 \u0441\u0435\u0440\
    \u0432\u0435\u0440"
  url: http://localhost:5001/